pyproject.toml
[project]
name = "sqlite-ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.135.1",
    "jinja2>=3.1.6",
    "langchain>=1.2.10",
    "langchain-openai>=1.1.11",
    "openai>=2.26.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.41.0",
]
